As discussed, throwing proper exceptions (recording the callstack the the throw-site) inside the carbon dll instead of opaque ints. At the border, catching them, storing in the handle table and returning corresponding handles as the return value. In the cpp layer, using the provided c-api functions to retrieve error details and re-throw as std::runtime_error.

Related work items: #1216258

+++ source/public/c_api/spxerror.h
@@ -7,10 +7,12 @@
 
 #pragma once
 
+#include <stdint.h>
+
 /// <summary>
 /// Type definition for SPX result codes.
 /// </summary>
-typedef unsigned int SPXHR;
+typedef uintptr_t SPXHR;
 
 /// <summary>
 /// Default result code indicating no error.
